pyproject.toml
[project]
name = "analytics-dbt-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dbt-core>=1.10.15",
    "dbt-snowflake>=1.10.2",
    "requests>=2.32.5",
    "snowflake-connector-python>=3.18.0",
]
